Move Banner, Deploy, Summary components to Contracts directory
