+++ packages/page-contracts/src/Contracts/Banner.tsx
@@ -7,7 +7,7 @@ import styled from 'styled-components';
 import { Button } from '@polkadot/react-components';
 import { useToggle } from '@polkadot/react-hooks';
 
-import { useTranslation } from './translate';
+import { useTranslation } from '../translate';
 
 interface Props {
   className?: string;
