+++ packages/page-contracts/src/Contracts/Summary.tsx
@@ -8,7 +8,7 @@ import { CardSummary, SummaryBox } from '@polkadot/react-components';
 import { useApi, useCall } from '@polkadot/react-hooks';
 import { formatNumber } from '@polkadot/util';
 
-import { useTranslation } from './translate';
+import { useTranslation } from '../translate';
 
 interface Props {
   trigger: number;
